clean up enchantment module code and use enchantment helper.

+++ src/main/java/net/machinemuse/powersuits/powermodule/special/SilkTouchModule.java
@@ -52,7 +52,7 @@ public class SilkTouchModule extends PowerModuleBase implements IEnchantmentModu
     }
 
     @Override
-    public short getLevel() {
+    public int getLevel() {
         return 1;
     }
 }
